moved factory inline in ref
https://backstage.io/docs/reference/backend-plugin-api.servicerefoptions
