+++ .changeset/soft-lands-doubt.md
@@ -1,0 +1,5 @@
+---
+'@roadiehq/backstage-plugin-argo-cd-backend': major
+---
+
+Removed requirement for external service factory include
