+++ plugins/backend/backstage-plugin-argo-cd-backend/src/index.ts
@@ -1,4 +1,3 @@
 export * from './service/router';
-export * from './factories/argocdService.factory';
 export * from './refs/argocdService.ref';
 export { ArgoService } from './service/argocd.service';
